requirements.txt
boto3>=1.28.0
fastmcp>=0.1.1
python-dotenv>=1.0.0
pytest>=6.2.5 